requirements.txt
flask>=3.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0

google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.90.0

# Real-time notifications
flask-socketio>=5.6.1
eventlet>=0.33.0